+++ test/eval-hacks/01-run-executable.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+echo "+------------------------------------------+"
+echo "|     I COULD HAVE WIPED OUT YOUR DISK     |"
+echo "+------------------------------------------+"
+touch "${file:-01-eval-hack.txt}"